Support for Parametrics subclasses of Parametrics<T>, T=T
Fixes to symantics for which enables variable basic T types of Parametrics objects.
toFloat/toDouble (deprecated) from core.
Added to<G,m> and toType<G> to replace toFloat/toDouble.
to<G,m> uses conversion operator.
Added conversion operator to ScalarPoint/Sphere.

+++ src/scene/selector/gmselector.c
@@ -58,7 +58,7 @@ namespace GMlib {
     _marked		= c.getInverse();
     _selected	= false;
     _root		= root;
-    translate( _position.toFloat() );
+    translate( _position );
     if(r != 1.0) scale(Vector<float,3>(r,r,r));
 
     _std_rep_visu = new VisualizerStdRep;
